+++ src/pas/plugins/authomatic/plugin.py
@@ -262,10 +262,24 @@ class AuthomaticPlugin(BasePlugin):
             if not userid:
                 logger.warn("None userid found. This should not happen!")
                 continue
-            if not userid.startswith(search_id):
-                continue
+
+            # search for a match in fullname, email and userid
             identity = self._useridentities_by_userid[userid]
+            search_term = search_id.lower()
             identity_userid = identity.userid
+            identity_fullname = identity.propertysheet.getProperty("fullname", "").lower()
+            identity_email = identity.propertysheet.getProperty("email", "").lower()
+            if (
+                not search_term in identity_userid
+                and not search_term in identity_fullname
+                and not search_term in identity_email
+            ):
+                continue
+
+#            if not userid.startswith(search_id):
+#                continue
+#            identity = self._useridentities_by_userid[userid]
+#            identity_userid = identity.userid
             ret.append(
                 {
                     "id": identity_userid,